Extract arguments into a trait
Lets us be a bit more DRY as we reuse this
