+++ core/src/main/groovy/org/grails/gorm/graphql/entity/operations/CustomOperation.groovy
@@ -5,13 +5,12 @@ import groovy.transform.CompileStatic
 import org.grails.datastore.mapping.model.MappingContext
 import org.grails.datastore.mapping.model.PersistentEntity
 import org.grails.gorm.graphql.GraphQLServiceManager
+import org.grails.gorm.graphql.entity.dsl.helpers.Arguable
 import org.grails.gorm.graphql.entity.dsl.helpers.Deprecatable
 import org.grails.gorm.graphql.entity.dsl.helpers.Describable
 import org.grails.gorm.graphql.entity.dsl.helpers.ExecutesClosures
 import org.grails.gorm.graphql.entity.dsl.helpers.Named
-import org.grails.gorm.graphql.entity.arguments.ComplexArgument
 import org.grails.gorm.graphql.entity.arguments.CustomArgument
-import org.grails.gorm.graphql.entity.arguments.SimpleArgument
 import org.grails.gorm.graphql.fetcher.interceptor.CustomMutationInterceptorInvoker
 import org.grails.gorm.graphql.fetcher.interceptor.CustomQueryInterceptorInvoker
 import org.grails.gorm.graphql.fetcher.interceptor.InterceptingDataFetcher
@@ -29,9 +28,8 @@ import static graphql.schema.GraphQLFieldDefinition.newFieldDefinition
  * @since 1.0.0
  */
 @CompileStatic
-abstract class CustomOperation<T> implements Named<T>, Describable<T>, Deprecatable<T>, ExecutesClosures {
+abstract class CustomOperation<T> implements Named<T>, Describable<T>, Deprecatable<T>, Arguable<T>, ExecutesClosures {
 
-    private List<CustomArgument> arguments = []
     private static InterceptorInvoker queryInvoker = new CustomQueryInterceptorInvoker()
     private static InterceptorInvoker mutationInvoker = new CustomMutationInterceptorInvoker()
     DataFetcher dataFetcher
@@ -43,54 +41,6 @@ abstract class CustomOperation<T> implements Named<T>, Describable<T>, Deprecata
     }
 
     OperationType operationType
-
-    private void handleArgumentClosure(CustomArgument argument, Closure closure) {
-        withDelegate(closure, argument)
-        argument.validate()
-        arguments.add(argument)
-    }
-
-    /**
-     * Creates an argument to the operation that is a list of a simple type.
-     * The list can not have more than 1 element and that element must be a class.
-     *
-     * @param name The name of the argument
-     * @param type The returnType of the argument
-     * @param closure To provide additional data about the argument
-     * @return The operation in order to chain method calls
-     */
-    CustomOperation argument(String name, List<Class> type, @DelegatesTo(value = SimpleArgument, strategy = Closure.DELEGATE_ONLY) Closure closure = null) {
-        CustomArgument argument = new SimpleArgument().name(name).returns(type)
-        handleArgumentClosure(argument, closure)
-        this
-    }
-
-    /**
-     * Creates an argument to the operation that is of the returnType provided.
-     *
-     * @param name The name of the argument
-     * @param type The returnType of the argument
-     * @param closure To provide additional data about the argument
-     * @return The operation in order to chain method calls
-     */
-    CustomOperation argument(String name, Class type, @DelegatesTo(value = SimpleArgument, strategy = Closure.DELEGATE_ONLY) Closure closure = null) {
-        CustomArgument argument = new SimpleArgument().name(name).returns(type)
-        handleArgumentClosure(argument, closure)
-        this
-    }
-
-    /**
-     * Creates an argument to the operation that is a custom type.
-     *
-     * @param name The name of the argument
-     * @param closure To provide additional data about the argument
-     * @return The operation in order to chain method calls
-     */
-    CustomOperation argument(String name, String typeName, @DelegatesTo(value = ComplexArgument, strategy = Closure.DELEGATE_ONLY) Closure closure) {
-        CustomArgument argument = new ComplexArgument().name(name).typeName(typeName)
-        handleArgumentClosure(argument, closure)
-        this
-    }
 
     /**
      * If the argument is true, the default list arguments created in the
